Fixed category title and bug fixed cards (position and size)
In this commit, I fixed the category pages' titles so they do not move out of sight when scrolling through the cards. I also added a media query that increases the size of the cards when in phone mode. I also repositioned the card-list section so that the first card on the list of cards is the first thing you see.